new data structures in progress
Former-commit-id: f4a271f01043f0c24d1d7ac9e5517c35d5b4194f

+++ branches/new_data_structures/GenPlay/src/edu/yu/einstein/genplay/dataStructure/scoredChromosomeWindow/ScoredChromosomeWindow.java
@@ -34,5 +34,5 @@ public interface ScoredChromosomeWindow extends ChromosomeWindow {
 	/**
 	 * @return the score of the window
 	 */
-	public double getScore ();
+	public float getScore ();
 }
